+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 
 ### Changed
 
+- Changed the message format.
 - Updated the dependencies.